# docs/swagger.yaml
basePath: /v1
definitions:
  model.IamUser:
    properties:
      namespace:
        type: string
      safe_id:
        type: string
      username:
        type: string
    required:
    - namespace
    - username
    type: object
  model.MigrateNamespace:
    properties:
      namespace:
        type: string
      safe_id:
        type: string
    required:
    - namespace
    type: object
  model.OnboardNamespace:
    properties:
      namespace:
        type: string
      safe_id:
        type: string
      username:
        type: string
    required:
    - namespace
    - username
    type: object
info:
  contact: {}
paths:
  /bird/onboard:
    post:
      consumes:
      - application/json
      description: In Dell ECS, creates a IAM user and an AccessKey. In Vault, stores
        the secret access keys and creates a JWT auth role bound to the Brid
      parameters:
      - description: the user to onboard
        in: body
        name: brid
        required: true
        schema:
          $ref: '#/definitions/model.IamUser'
      produces:
      - application/json
      responses: {}
      summary: onboard a brid to a namespace as IAM user
      tags:
      - Brid
  /iamuser/{username}:
    delete:
      description: In Dell ECS, deletes the IAM user. In Vault, deletes the role and
        the JWT auth role (if any)
      parameters:
      - description: the user to delete
        in: path
        name: brid
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: delete a IAM user
      tags:
      - IamUser
  /iamuser/onboard:
    post:
      consumes:
      - application/json
      description: In Dell ECS, creates a IAM user and an AccessKey. In Vault, stores
        the secret access keys and creates a JWT auth role bound to the Brid
      parameters:
      - description: the user to onboard
        in: body
        name: brid
        required: true
        schema:
          $ref: '#/definitions/model.IamUser'
      produces:
      - application/json
      responses: {}
      summary: onboard a IAM user in a namespace
      tags:
      - IamUser
  /namespace/migrate:
    post:
      consumes:
      - application/json
      description: In Dell ECS, creates a IAM user (and AccessKey) for the Native
        users, creates a second AccessKey for existing IAM users. In Vault, stores
        the Secret Access Keys
      parameters:
      - description: the namespace to migrate
        in: body
        name: ns
        required: true
        schema:
          $ref: '#/definitions/model.MigrateNamespace'
      produces:
      - application/json
      responses: {}
      summary: migrates a namespace
      tags:
      - Namespace
  /namespace/onboard:
    post:
      consumes:
      - application/json
      description: In Dell ECS, creates a namespace, a IAM user and an AccessKey.
        In Vault, stores the Secret Access Key
      parameters:
      - description: the namespace to onboard
        in: body
        name: ns
        required: true
        schema:
          $ref: '#/definitions/model.OnboardNamespace'
      produces:
      - application/json
      responses: {}
      summary: onboards a namespace
      tags:
      - Namespace
swagger: "2.0"